requirements.txt
line-bot-sdk==2.3.0
openai==0.26.0
fastapi[all]
